+++ packages/DataX/es/checker.js
@@ -1,5 +1,4 @@
-import TypeChecker, { hasOwn } from "@lihzsky/type-checker";
-TypeChecker.extend('isReservedProperty', function (value) { return !/^__bean_/.test(value); });
-TypeChecker.extend('hasOwn', hasOwn);
+import TypeChecker from "@lihzsky/type-checker";
+TypeChecker.isReservedProperty = function (value) { return !/^__bean_/.test(value); };
 export default TypeChecker;
 //# sourceMappingURL=checker.js.map