+++ packages/DataX/src/checker.ts
@@ -1,6 +1,5 @@
-import TypeChecker, { hasOwn } from "@lihzsky/type-checker"
+import TypeChecker from "@lihzsky/type-checker"
 
-TypeChecker.extend('isReservedProperty', (value: string) => !/^__bean_/.test(value))
-TypeChecker.extend('hasOwn', hasOwn)
+TypeChecker.isReservedProperty = (value: string) => !/^__bean_/.test(value)
 
 export default TypeChecker