+++ packages/DataX/CHANGELOG.md
@@ -1,6 +1,20 @@
 # @lihzsky/data-x
 
+## 1.1.2
+
+### Patch Changes
+
+- 依赖包更新
+
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @lihzsky/type-checker@1.0.0
+
 ## 1.1.0
+
 ### Minor Changes
 
 - 优化转换方法
@@ -11,31 +25,37 @@
   - @lihzsky/type-checker@0.0.2
 
 ## 1.0.0
+
 ### Major Changes
 
 - 重构转化类
 
 ## 0.0.5
+
 ### Patch Changes
 
 - 修复打包文件未更新问题
 
 ## 0.0.4
+
 ### Patch Changes
 
-- 修复data为null的情况
+- 修复 data 为 null 的情况
 
 ## 0.0.3
+
 ### Patch Changes
 
 - 55a8d2c: 优化嵌套结构解析
 
 ## 0.0.2
+
 ### Patch Changes
 
-- 修复isSameType is not function
+- 修复 isSameType is not function
 
 ## 0.0.1
+
 ### Patch Changes
 
 - 创建新版本