+++ packages/TypeChecker/CHANGELOG.md
@@ -1,11 +1,19 @@
 # @lihzsky/type-checker
 
+## 1.0.0
+
+### Major Changes
+
+- 修改 TypeChecker 实现方案
+
 ## 0.0.2
+
 ### Patch Changes
 
 - 优化转换方法
 
 ## 0.0.1
+
 ### Patch Changes
 
 - 创建新版本