+++ packages/UtilShared/CHANGELOG.md
@@ -1,12 +1,27 @@
 # @lihzsky/util-shared
 
+## 0.0.4
+
+### Patch Changes
+
+- 依赖包更新
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @lihzsky/type-checker@1.0.0
+
 ## 0.0.2
+
 ### Patch Changes
 
 - Updated dependencies
   - @lihzsky/type-checker@0.0.2
 
 ## 0.0.1
+
 ### Patch Changes
 
 - 新版本发布